chore(lint): fix file_name violations (#279)
Rename or split 14 files so each matches a type/extension declared in
the file, per the SwiftLint `file_name` rule. Where a file held two
unrelated views or a pile of independent test suites
(`EarmarkFormSheet`, `AutomationServiceTests`) it is split into one
file per subject — matching the existing `AccountStore*Tests.swift`
pattern — rather than renamed to match only the first type.

Stacked on #361: its RecordMapping and AnalysisStoreTests splits
already cleared those two baseline entries, so this PR only handles
the remaining 14.

Baseline regenerated; file_name drops from 16 to 0 and no other rule
count increases.

Fixes #279
