+++ Automation/AppleScript/Commands/NSScriptCommand+ScriptHelpers.swift
